+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 - Fix entity event to var causal chain sometimes not working on var increments
 - Fix most crashes on remote cutscene skip (there are still some rare ones)
+- Fix ig.ENTITY.WallVertical & ig.ENTITY.WallHorizontal initial remote state
 
 ## [0.5.7] 2025-11-10